# interfaces/rest/openapi.yaml
openapi: 3.0.3
info:
  title: EventFlow REST API
  description: |
    RESTful API for EventFlow neuromorphic computing platform.

    Provides programmatic access to neuromorphic model validation, execution,
    tracing, and management capabilities through standard HTTP endpoints.
  version: 0.1.0
  contact:
    name: EventFlow Team
    url: https://github.com/eventflow/eventflow
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.eventflow.dev/v1
    description: Production API
paths:
...
  /validate/eir:
    post:
      summary: Validate EIR document
      description: Validates an EventFlow Intermediate Representation JSON document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: EIR JSON document to validate
      responses:
        '200':
          description: Validation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
  /models:
    get:
      summary: List available models
      description: Returns list of available neuromorphic models
      parameters:
        - name: profile
          in: query
          schema:
            type: string
            enum: [BASE, REALTIME, LEARNING, LOWPOWER]
        - name: domain
          in: query
          schema:
            type: string
            description: Filter by application domain
      responses:
        '200':
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelSummary'

    post:
      summary: Upload model
      description: Upload and register a new neuromorphic model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelUpload'
      responses:
        '201':
          description: Model uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
...
  /execute/{model_id}/run:
    post:
      summary: Execute model
      description: Runs a model with provided inputs and generates traces
      parameters:
        - name: model_id
          in: path
          required: true
          schema:
            type: string
        - name: backend
          in: query
          schema:
            type: string
            default: cpu-sim
        - name: async
          in: query
          schema:
            type: boolean
            default: false
            description: Execute asynchronously
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionRequest'
      responses:
        '200':
          description: Execution completed synchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResult'
        '202':
          description: Execution started asynchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionJob'
        '404':
          description: Model not found
...
components:
  schemas:
...
    ValidationResponse:
      type: object
      required: [valid, issues]
      properties:
        valid:
          type: boolean
        issues:
          type: array
          items:
            type: object
            properties:
              severity:
                type: string
                enum: [error, warning, info]
              code:
                type: string
              message:
                type: string
              location:
                type: object
                properties:
                  path:
                    type: string
                  line:
                    type: integer
                  column:
                    type: integer
    ValidationError:
      type: object
      required: [error, details]
      properties:
        error:
          type: string
        details:
          $ref: '#/components/schemas/ValidationResponse'
    ModelSummary:
      type: object
      required: [id, name, version]
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        profile:
          type: string
          enum: [BASE, REALTIME, LEARNING, LOWPOWER]
        domain:
          type: string
        created_at:
          type: string
          format: date-time
    Model:
      allOf:
        - $ref: '#/components/schemas/ModelSummary'
        - type: object
          properties:
            description:
              type: string
            author:
              type: string
            license:
              type: string
            eir_size_bytes:
              type: integer
            capabilities:
              type: array
              items:
                type: string
    ModelUpload:
      type: object
      required: [name, eir]
      properties:
        name:
          type: string
        version:
          type: string
          default: "1.0.0"
        description:
          type: string
        profile:
          type: string
          enum: [BASE, REALTIME, LEARNING, LOWPOWER]
          default: BASE
        domain:
          type: string
        eir:
          type: object
          description: EIR JSON document
        golden_trace:
          type: object
          description: Optional golden trace data
...
    ExecutionRequest:
      type: object
      required: [inputs]
      properties:
        inputs:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              uri:
                type: string
              format:
                type: string
                enum: [json, jsonl]
              data:
                type: object
        plan_id:
          type: string
          description: Pre-built plan ID to use
        parameters:
          type: object
          description: Execution parameters
    ExecutionResult:
      type: object
      required: [job_id, status, traces]
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [completed, failed]
        traces:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              format:
                type: string
              url:
                type: string
        execution_time_us:
          type: integer
        error:
          type: string
    ExecutionJob:
      type: object
      required: [job_id, status, created_at]
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        created_at:
          type: string
          format: date-time
        progress_percent:
          type: number
          minimum: 0
          maximum: 100
        estimated_completion:
          type: string
          format: date-time
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
security:
  - ApiKeyAuth: []
  - BearerAuth: []